refactor: improve service registration and facade implementation
- Replace fully qualified class names with imported Storage facade
- Update facade docblock with new method signatures
- Change service registration to use string identifiers
- Add Override attribute to facade and service provider methods
- Simplify test command by using imported facades

+++ src/Services/CacheFailoverManager.php
@@ -215,7 +215,7 @@ class CacheFailoverManager
                 $retrieved = Cache::store($store)->get($testKey);
                 Cache::store($store)->forget($testKey);
 
-                $responseTime = (microtime(true) - $startTime) * 1000;
+                $responseTime = (float) ((microtime(true) - $startTime) * 1000);
 
                 if ($retrieved === $testValue) {
                     $results[$name] = [
